move function process to another file
